requirements.txt
anthropic>=0.42.0
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests>=2.32.0
python-dotenv>=1.0.0
schedule>=1.2.0
